shmoverride: do not remove shm.id file if created by other instance
If second instance of X server is started, shm.id creation will fail.
But in such a case, on X server exit, do not remove it, as it belong to
the other instance. Otherwise it will break GUI in that one.

Fixes QubesOS/qubes-issues#2171

+++ shmoverride/shmoverride.c
@@ -42,7 +42,7 @@ static int (*real_shmdt) (const void *shmaddr);
 static int (*real_shmctl) (int shmid, int cmd, struct shmid_ds * buf);
 
 static int local_shmid = 0xabcdef;
-static struct shm_cmd *cmd_pages;
+static struct shm_cmd *cmd_pages = NULL;
 static struct genlist *addr_list;
 #ifdef XENCTRL_HAS_XC_INTERFACE
 static xc_interface *xc_hnd;
@@ -177,8 +177,10 @@ int __attribute__ ((constructor)) initfunc()
 
 int __attribute__ ((destructor)) descfunc()
 {
-    real_shmdt(cmd_pages);
-    real_shmctl(local_shmid, IPC_RMID, 0);
-    unlink(SHMID_FILENAME);
+    if (cmd_pages) {
+        real_shmdt(cmd_pages);
+        real_shmctl(local_shmid, IPC_RMID, 0);
+        unlink(SHMID_FILENAME);
+    }
     return 0;
 }
